implements the zeromorph polynomial commitment scheme

+++ spartan/src/product_tree.rs
@@ -38,7 +38,7 @@ impl<F: PrimeField> ProductCircuit<F> {
     let mut left_vec: Vec<DensePolynomial<F>> = Vec::new();
     let mut right_vec: Vec<DensePolynomial<F>> = Vec::new();
 
-    let num_layers = poly.len().log_2() as usize;
+    let num_layers = poly.len().log_2();
     let (outp_left, outp_right) = poly.split(poly.len() / 2);
 
     left_vec.push(outp_left);
@@ -193,7 +193,7 @@ impl<F: PrimeField> ProductCircuitEvalProof<F> {
       let mut poly_C = DensePolynomial::new(EqPolynomial::new(rand.clone()).evals());
       assert_eq!(poly_C.len(), len / 2);
 
-      let num_rounds_prod = poly_C.len().log_2() as usize;
+      let num_rounds_prod = poly_C.len().log_2();
       let comb_func_prod = |poly_A_comp: &F, poly_B_comp: &F, poly_C_comp: &F| -> F {
         *poly_A_comp * *poly_B_comp * *poly_C_comp
       };
@@ -241,7 +241,7 @@ impl<F: PrimeField> ProductCircuitEvalProof<F> {
   where
     G: CurveGroup<ScalarField = F>,
   {
-    let num_layers = len.log_2() as usize;
+    let num_layers = len.log_2();
     let mut claim = eval;
     let mut rand: Vec<F> = Vec::new();
     //let mut num_rounds = 0;
@@ -308,7 +308,7 @@ impl<F: PrimeField> ProductCircuitEvalProofBatched<F> {
       let mut poly_C_par = DensePolynomial::new(EqPolynomial::<F>::new(rand.clone()).evals());
       assert_eq!(poly_C_par.len(), len / 2);
 
-      let num_rounds_prod = poly_C_par.len().log_2() as usize;
+      let num_rounds_prod = poly_C_par.len().log_2();
       let comb_func_prod = |poly_A_comp: &F, poly_B_comp: &F, poly_C_comp: &F| -> F {
         *poly_A_comp * *poly_B_comp * *poly_C_comp
       };
@@ -448,7 +448,7 @@ impl<F: PrimeField> ProductCircuitEvalProofBatched<F> {
   where
     G: CurveGroup<ScalarField = F>,
   {
-    let num_layers = len.log_2() as usize;
+    let num_layers = len.log_2();
     let mut rand: Vec<F> = Vec::new();
     //let mut num_rounds = 0;
     assert_eq!(self.proof.len(), num_layers);
